Start on test for deleting old transactions
