+++ java/statestore/src/test/java/sleeper/statestore/transactionlog/TransactionLogStateStoreSnapshotsIT.java
@@ -25,12 +25,8 @@ import sleeper.core.schema.type.StringType;
 import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.FileReferenceFactory;
 import sleeper.core.statestore.StateStore;
-import sleeper.core.statestore.transactionlog.InMemoryTransactionLogSnapshotSetup;
-import sleeper.core.statestore.transactionlog.InMemoryTransactionLogSnapshotSetup.SetupStateStore;
-import sleeper.core.statestore.transactionlog.TransactionLogStateStore;
 
 import java.util.List;
-import java.util.function.Consumer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.configuration.properties.table.TableProperty.ADD_TRANSACTION_MAX_ATTEMPTS;
@@ -38,7 +34,6 @@ import static sleeper.configuration.properties.table.TableProperty.TIME_BETWEEN_
 import static sleeper.configuration.properties.table.TableProperty.TIME_BETWEEN_TRANSACTION_CHECKS_MS;
 import static sleeper.core.schema.SchemaTestHelper.schemaWithKey;
 import static sleeper.core.statestore.FileReferenceTestData.DEFAULT_UPDATE_TIME;
-import static sleeper.core.statestore.transactionlog.InMemoryTransactionLogSnapshotSetup.setupSnapshotWithFreshState;
 
 public class TransactionLogStateStoreSnapshotsIT extends TransactionLogStateStoreOneTableTestBase {
 
@@ -159,23 +154,8 @@ public class TransactionLogStateStoreSnapshotsIT extends TransactionLogStateStor
         });
     }
 
-    protected StateStore stateStore(Consumer<TransactionLogStateStore.Builder> config) {
-        TransactionLogStateStore.Builder builder = stateStoreBuilder(tableProperties);
-        config.accept(builder);
-        return stateStore(builder);
-    }
-
     private FileReferenceFactory fileFactory() {
         return FileReferenceFactory.fromUpdatedAt(partitions.buildTree(), DEFAULT_UPDATE_TIME);
-    }
-
-    protected void createSnapshotWithFreshStateAtTransactionNumber(
-            long transactionNumber, SetupStateStore setupState) throws Exception {
-        InMemoryTransactionLogSnapshotSetup snapshotSetup = setupSnapshotWithFreshState(
-                tableProperties.getStatus(), tableProperties.getSchema(), setupState);
-        DynamoDBTransactionLogSnapshotStore snapshotStore = new DynamoDBTransactionLogSnapshotStore(instanceProperties, tableProperties, dynamoDBClient, configuration);
-        snapshotStore.saveFilesSnapshot(snapshotSetup.createFilesSnapshot(transactionNumber));
-        snapshotStore.savePartitionsSnapshot(snapshotSetup.createPartitionsSnapshot(transactionNumber));
     }
 
 }
